// src/api/__generated__/wallet/model/notificationUnreadCountResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 站内通知未读计数
 */
export interface NotificationUnreadCountResponse {
  /** 收益 Tab 未读数 */
  incomeUnread?: string;
  /** 互动 Tab 未读数 */
  interactionUnread?: string;
  /** 总未读数 */
  totalUnread?: string;
}