// src/api/__generated__/wallet/notification/notification.ts
export type listResponse200 = {
  data: ResultCursorPageResponseNotificationItemResponse;
  status: 200;
};

export type listResponse400 = {
  data: ResultObject;
  status: 400;
};

export type listResponse405 = {
  data: ResultObject;
  status: 405;
};

export type listResponse500 = {
  data: ResultObject;
  status: 500;
};

export type listResponseSuccess = listResponse200 & {
  headers: Headers;
};
export type listResponseError = (
  | listResponse400
  | listResponse405
  | listResponse500
) & {
  headers: Headers;
};

export type listResponse = listResponseSuccess | listResponseError;